refactor(server): migrate autoload from src/ to server/src/
- Update composer.json PSR-4 autoload to use server/src/
- Update server.php to use NimbusIO\Core\GameEngine
- Remove legacy src/ directory (duplicate code)
- Change Wizard hp/maxHp from int to float to fix precision warning

+++ server/src/Entities/Wizard.php
@@ -17,8 +17,8 @@ class Wizard
     public float $velX = 0;
     public float $velY = 0;
     public float $angle = 0;
-    public int $hp;
-    public int $maxHp = 100;
+    public float $hp;
+    public float $maxHp = 100;
     public float $mana;
     public float $maxMana = 100;
     public int $xp = 0;
